Namespace the version file on the file system
The version file is at the toplevel dir in `lib`, which makes it not
namespaced on the file system. Thus `require "version"` is really
requiring `dotenv_validator`'s version, despite the fact that this is
unexpected. The solution that people have arrived at is to namespace
files within a directory that matches the gem name. So, this movew
`lib/version.rb` to `lib/dotenv_validator/version.rb`.

For a bit more context on this, I go into it
[here](https://github.com/JoshCheek/elective-building-a-gem/blob/085283ee848287f2b57f8ec5da963c5d775eb3de/Day2.md#namespacing-our-gems-files-and-classes).

```sh
gem install --no-document dotenv_validator -v 1.0.0 >/dev/null && ruby -e '
   features = $LOADED_FEATURES.dup
   require "version"
   puts $LOADED_FEATURES - features'
/Users/josh/.rubies/ruby-3.0.2/lib/ruby/3.0.0/rubygems/bundler_version_finder.rb
/Users/josh/.gem/ruby/3.0.2/gems/dotenv_validator-1.0.0/lib/version.rb
```